% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as.data.frame.hm_result}
\alias{as.data.frame.hm_result}
\title{Convert an \code{hm_result} to data frame}
\usage{
\method{as.data.frame}{hm_result}(x, ...)
}
\arguments{
\item{x}{An \code{hm_result} object.}

\item{...}{Unused.}
}
\value{
A plain data frame.
}
\description{
Convert an \code{hm_result} to data frame
}
