% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{hm_result}
\alias{hm_result}
\title{Create an \code{hm_result} object}
\usage{
hm_result(x)
}
\arguments{
\item{x}{A data frame of metric results.}
}
\value{
An object with class \code{hm_result}.
}
\description{
Create an \code{hm_result} object
}
