% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.hm_result}
\alias{print.hm_result}
\title{Print an \code{hm_result}}
\usage{
\method{print}{hm_result}(x, ...)
}
\arguments{
\item{x}{An \code{hm_result} object.}

\item{...}{Additional arguments passed to \code{\link[=print]{print()}}.}
}
\value{
The input object, invisibly.
}
\description{
Stable S3 print method for objects created by \code{\link[=hm_result]{hm_result()}}.
}
